Added logging for contract debugging

+++ contracts/AdditionalZkBNB.sol
@@ -566,7 +566,7 @@ contract AdditionalZkBNB is Storage, Config, Events, ReentrancyGuard, IERC721Rec
         TxTypes.TxType priorReqType = priorityRequests[_priorityRequestId].txType;
         // incorrect priority _tx type
         if (priorReqType != TxTypes.TxType.FullExit) {
-            console.log("ERROR! priorReqId=%s,priorReqType=%s,required=FullExit", _priorityRequestId, priorReqType);
+            console.log("ERROR! priorReqId=%s,priorReqType=%s,required=FullExit", _priorityRequestId, uint(priorReqType));
         }
         require(priorReqType == TxTypes.TxType.FullExit, "J");
